window-group: Subtract the unredirected window in stage space
meta_window_group_paint tries to carefully figure out which parts of the
scene it can avoid painting. One area it avoids painting is the region
of the screen occupied by an unredirected window (if there's one present).

In subtracting that region, it gets the coordinate spaces confused, and
ends up subtracting the area at the wrong offset.

This commit translates the region to the correct offset before
subtracting.

https://bugzilla.gnome.org/show_bug.cgi?id=677116