Ensure steam notifications appear over the panel. Fixes #2357
I'd prefer to simply move the popup into the work area, but for
for some reason, I can't apply our tooltip fix to these popups -
with tooltips, we check if they're overlapping and move them to a
better spot.  These popups I cannot seem to move.

It seems like Steam is using the monitor geometry to calculate the
position for this popup.  I think it would be more appropriate to
get the primary monitor's *work area* which is the monitor geometry
minus any struts (panels, other items to be considered part of the
DE's chrome) - this would return the usable workarea where any
calculation should take place within.  Note that doing this could
end up eliminating the need for DE hacks in Steam's code, as well
as ours.