frames: Fix clicking maximize button from a top- or bottom-tiled state -
Whether clicking this button will 'maximize' or 'unmaximize' was only
being determined by the window frame style - if the window was styled as
'maximized' already (as it is when it's top- or bottom- tiled), muffin
was assuming you wanted to 'unmaximize' the window, which is incorrect -
always maximize when clicking the maximize button from a tiled state.

Ideally, we would have top-tiled and bottom-tiled frame styles, but as
that would require every metacity theme to add this support (when most
only have left, right, and maximize styles,) we co-opted the maximize
style when tiling top and bottom.