Add _NET_WM_STATE_MAXIMIZED_VERT to the _NET_WM_STATE for tiled windows.
This re-adds most of what was reverted in 7d6d49e98020ba67ce21ba3e, which
fixed an issue where a tiled window would re-open in a maximized state.

The original commit (6f67cf3cd822111041756dd6f) was to trick gtk into
hiding the GtkWindow resize grip when windows were tiled.

It turns out that gtk checks _NET_WM_STATE_MAXIMIZED_VERT alone to determine
if a window should be decorated (c-s-d) as tiled or not.  Without that atom
set, it seems that some empty padding is added around a CSD window, which
should be removed during maximization or tiling.  I may be mistaken in this
in some way, but the net result is that we end up receiving a client window
size that is larger than the visible area of that window (xwindow and CSD
frame included.)

Because only _VERT is required to achieve this, instead of both _HORZ and
_VERT, bug #151 is still happily fixed.

Fixes linuxmint/Cinnamon#3667