Remove kill-switch-workspace signal
The kill-switch-workspace signal is used when before switch-workspace
signal is fired for the existing switch animation to be cancelled.
However, this mechanism is not ideal since the switch-workspace and
kill-switch-workspace events are separated.

Using the old model, say we switch to left workspace, and during that
animation, we press Ctrl-Alt-Right to switch back to the right
workspace. What would happen is that we will finish the animation early,
land on the left workspace, and start a new animation.

By letting the wm handle the killing work when receiving the
switch-workspace, we can do better. We can immediately halt the windows'
movement, and let them animate to where Ctrl-Alt-Right would bring them
to, from the position where they started, which would look smoother.

This new way of doing things is already implemented in previous commits,
and this commit removes the kill-switch-workspace signal completely.