meta-window-actor.c: Only override the unredirect heuristic with our unredirect-fullscreen-windows preference.  This is based on a few assumptions:
- If a window explicitly requests that we bypass compositing,
  we should assume it knows what it's doing, and let it.
- The same holds true for with override_redirect windows.
- The preference was only added because the heuristic had been
  added.  The other checks already did or would exist.

The problem with the current behavior is the preference effectively
short-circuits determining how to handle a window based on whether
the setting is TRUE or FALSE.  This prevents GtkWindowType.POPUP
windows (which are OR) from becoming 'fullscreen', which is fully
supported in the rest of muffin.  In turn, this prevents Cinnamon
from detecting that a window is indeed fullscreen, and turning off
the pick/paint of our stage elements like panels and desklets.  This
causes those elements to eat any mouse or click events that take place
over them, even if they're obscured by an OR window (due to Cinnamon's
'captured-event' callback in main.js).  The detection of fullscreen
windows is handled in layout.js (connecting to MetaScreen's
'in-fullscreen-changed' signal.)