Make background transitions configurable
This is needed for compatibility reasons.

In LMDE 2 Betsy, when running in Virtualbox, blending doesn't work well.
It's as if the opacity of the bottom and top actors were added and the resulting
picture looks too shiny/clear as if its gamma was pushed too high.

This commits makes the transition configurable, so that we can ship
with a simple fade-in effect in LMDE instead of the blend.