theme: Draw window borders with the gtk3 theme
This drops all support for the metacity theme format. Borders will now drawn
by the gtk theme. This is a rough implementation still and some decisions need
to be made with regards to handling things like fringe button support.

Thanks to the upstream mutter devs who did all the original work this is
based on.