window.c: Change behavior of meta_window_get_all_monitors() to return only those monitors of a fullscreen window that are completely covered by the window.
When using Kodi with multiple monitors, it seems to start full screen,
but at a slightly larger size than its monitor, causing the other
monitor to hide its panels as well.  This can be corrected by toggling
Kodi out of and back into fullscreen mode, but a 10px wide strip of a
window should not constitute 'fullscreen' on that monitor, just because
the window is fullscreen on another.

Ref: linuxmint/Cinnamon#5460