window: Move legacy fullscreen workaround out of constraints and into meta_window_move_resize_request.
In constraints, it was catching Firefox in its new client-decorated
mode, and causing the window to fullscreen at incorrect times such
as during a drag from a maximized state.  Adding a check there to
subtract the frame (and thus make sure a framed window never gets
fullscreened) won't work because at least for the firefox case, during
a drag there is no real border to work with.

Moving it to its new spot actually keeps the problem case here from
ever reaching that code, but it can be seen operating with other windows
during tile operations.  To be honest, we'd never really seen this code
trigger previous to this, so I don't think there's an issue either way
with this fix.

ref: https://bugzilla.gnome.org/show_bug.cgi?id=730681