window.c: set the "saved_maximize" internal state based on the resulting state of the window after a maximize call, not the directions that were requested for the call.
For normal, wm-decorated windows, our titlebar buttons and actions correctly
combine maximize_horizontal and _vertical in one operation here, as do
keyboard shortcuts, regardless of window type.  It seems that csd windows
send separate requests for maximizing - first one direction then the other,
so our check was ignoring those windows, when maximized using the client
controls or titlebar action.

Instead, check the state of the window that the call would result in.  If
we end up with a fully maximized window, we save that state.

The "saved_maximize" state is used when a window is dragged loose from its
maximized state, and is primed to be able to be re-maximized on the same
or another monitor, even when the normal tile action for the upper edge is
not maximize.