window.c: Always use the full window extents (including invisible) when running meta_window_get_all_monitors() on CSD windows.
When a window is requested to be made fullscreen, this method gets called
from screen.c as a result of meta_screen_queue_fullscreen_check() during
the internal fullscreen process.  For CSD windows, our call here to
meta_window_get_outer_rect() offsets inward (smaller) for the window's
custom frame extents, so we only end up with the visible rectangle.
Since the monitor fullscreen check gets called *before* this extra padding
is removed by another update from gtk, the window fails the fullscreen
check, because it is not monitor-sized.

For CSD windows, we'll get the raw window rect, which will always be the
extreme outer window size.

Fixes #7614