meta-window-actor.c: Don't shadow dock type windows as a rule.
In testing 9ce812547b30aa5af8e, most 'dock' programs didn't actually
use the dock window-type hint, and would instead be caught by the
opacity check.  We should, however, as a rule, skip shadowing these,
as some programs do actually use the hint, and are expected not to
have shadows as a result.