window-props: Start respecting the _GTK_HIDE_TITLEBAR_WHEN_MAXIMIZED window hint.
This was removed early on primarily because upstream programs for
mint were being patched to work better with the whole 'global menu'
craze, without regard for whether or not a global menu of any sort
was being used.

Add a user pref 'just in case' someone encounters a program that
abuses this feature.

This does not affect gtk client-side-decorated windows.

Fixes linuxmint/Cinnamon#7681