Re-add gsettings key org.cinnamon.muffin button-layout
This key is obsolete and doesn't do anything since Cinnamon 3.8.

In Debian Stretch, Chromium relies on it and crashes if the key
is absent. Debian won't fix anything that doesn't directly affect
them and they're running an old version of Cinnamon (where the key
is present). So we're fixing it here for them instead.