Don't set the mask texture on every paint
This makes it so the mask texture only updates on shape region change
and size change, or specific conditions that self-decorated windows need
to not have artifacts, such as maximizing and unmaximizing. Mask texture
generation seems to be responsible for a lot of the compositor latency
because Cinnamon performs much better when their code paths are disabled
(with artifacts around every window).