window-actor: Phase out the size/position changed signals, and connect to shaped texture's size changed signal
MetaWindowActor will still emit its own size/position signals when
responding to MetaShapedTexture and MetaWindow respectively, so
everything doesn't break. This will give Cinnamon and xlets time to
switch to connecting to MetaWindow instead.

Adapted from
https://github.com/GNOME/mutter/commit/a0ef7c7142557913077b398cc219665a21650b95
https://github.com/GNOME/mutter/commit/9a6a189e3697fbcfbe4def9d9178d2b1db2ee9e3

In Mutter, this is essentially what happens - SurfaceActor serves as a
middle man between MetaWindowActor and MetaShapedTexture. This can be
worked into a port of SurfaceActor if there is a need later.