Fix VLC fullscreen controller position
VLC fullscreen controller's window type was changed in commit
https://git.videolan.org/?p=vlc.git;a=commit;h=d60312d3ad4c4e27f3455b4f579999b826d3948c
Since then, it got pushed up by the height of a bottom panel, so that is
was not displayed at the bottom monitor edge.
This fixes it.
This will also affect other "normal" windows that have a transient
parent in fullscreen.