window-actor: Create mask textures less often on idle
By waiting for the next idle, we are guaranteed to have all the geometry
updates for a window actor complete, so when the mask texture is
generated, we're basing it off of accurate window dimensions. This
allows us to call ensure_mask less frequently, and resolves the need for
previous work-arounds.