window-actor: Fix corners not updating on size change (#394)
* window-actor: Create mask textures less often on idle

By waiting for the next idle, we are guaranteed to have all the geometry
updates for a window actor complete, so when the mask texture is
generated, we're basing it off of accurate window dimensions. This
allows us to call ensure_mask less frequently, and resolves the need for
previous work-arounds.

* window: Add meta_window_get_client_area_rect

* Fix work-arounds, don't calculate borders on all windows

The resizing signal work-around in MetaWindow is not needed - the
correct fix is calling update_corners before creating the mask texture
in check_needs_reshape because the overlay path gets set in
MetaShapedTexture in update_corners, and the mask generation depends on
that.

We don't need to calculate MetaFrameBorders on frameless windows - this
call is heavy, even with the caching added in
https://github.com/linuxmint/muffin/commit/8779c27f78150970b98e5681acfb5b792d8fea35.
Use the client area rect in check_needs_reshape, and then pull the bits
out of meta_window_get_outer_rect that we need in update_corners for
framed windows. This was calling meta_frame_calc_borders a second time
on a single paint, and is not used here in #371.

Also changes g_idle_add_full to meta_later_add because the delay is only
needed on size change when the window geometry hasn't been computed yet.
We need the mask texture to render before the next redraw, otherwise a
performance regression occurs.

* Clean up changes

Creating the mask texture on idle caused frame pacing problems, and
unreliably caused artifacts on frameless chromium windows. The fix for
framed window corner artifacts was moving update_corners before
meta_window_actor_reset_mask_texture.