Add back the icon creation functions, fix blurry icons
Removing these prevents us from getting icons for some window backed
apps, such as Melange - which could solve this by using XAppGtkWindow,
but this could pose an issue for third party apps that have window
backed windows.

To fix the blurry icons, we are going to convert
meta_window_update_icon_now to meta_window_set_icon, which will accept
any size that will be passed in from `window_backed_app_get_icon` in
Cinnamon. We are not going to create icons for every window anymore, and
instead let Cinnamon use this function selectively when it needs to.

We can still remove the mini icon, and all of the theme/frame processing
because we don't set the icon in the title bar.