Only call meta_frames_calc_geometry on frame manage and size change
Before, functions received MetaFrameGeometry by re-calculating it every
time, but it really will never change unless the window rect changes, as
that is what is passed to meta_theme_calc_geometry. It is an expensive
function.

The first attempt at optimizing this was checking the rect whenever the
calculation function was called, and storing a single MetaFrameGeometry
object on the MetaFrames struct, and returning the cache on no size
change. That was incorrect because MetaFrameGeometry is shared by all
windows. As a result, artifacts were occurring. When simply moving the
cache to MetaUIFrame, this caused an increase in latency.

The new approach is to only call meta_frames_calc_geometry in response
to size change, and to replace previous places where
meta_frames_calc_geometry is called with the cached object on
MetaUIFrame. This ensures the geometry is only calculated when it needs
to be, and that it stays current.