Merge MetaShapedTexture into MetaWindowActor
This is being done because MetaShapedTexture is painting entirely with
Cogl and Cairo - there's nothing requiring it to be a child of
MetaWindowActor, or to be a Clutter specific instance. Clutter is just the mechanism used to integrate the compositor into the UI layer, and structure its life cycle methods. The more actors involved in rendering a single window, the more things that can go wrong - like windows flickering from broken actor synchronization, which I suspect is related.