screen: Fix segfault when restarting Cinnamon on a workspace other than the first
This was occurring in meta_compositor_switch_workspace, a call that
originates from meta_screen_new. This is caused by the removal of
MetaCompScreen, where it bailed anyway after a NULL check. The
compositor is not actually initialized at this point, so we shouldn't be
using any methods that propagate to the compositor in meta_screen_new.

I tested moving this code to compositor.c in
meta_compositor_manage_screen, but it didn't change the initial
workspace. Since it doesn't work, it is being removed.