window-actor: Optimize rendering for completely obscured windows
Currently the visible property only stays in sync with whether or not
the actor is visible, but doesn't take into account if the window is
unminimized but completely obscured by another window on top. This
reduces the process time spent on these windows.

- Checks if the actor is obscured when the stack is synchronized, not on
every paint, and caches the value.
- Various paint functions will return early if a window is obscured.
- In order to not break live thumbnail previews, a method is exposed to
GI that will override the optimization, but it is the caller's
responsbility to make sure that it toggles the override off (obscured ==
TRUE). The previous obscured state will be cached and restored if true
is passed to meta_window_actor_set_obscured. All widgets using live
thumbnails will need to call this function when showing a window
texture, so obscured windows are not treated like minimized ones (frozen
texture).