Toggle the NO_LAYOUT actor flag on window_group before and after grab operations
Adding this flag to the window group was originally intended to help
improve window dragging smoothness, but its benefit there seems
negligible. However, it does improve input latency with window
rendering, so it's worth
finding a way to make it work without artifacts.

This disables the NO_LAYOUT flag during window dragging, which was
causing windows to flicker in and out of view.