Always clear the icon cache
Now that icons are initialized from Cinnamon, we need to ensure an icon
is always created for the specified size. The texture is copied to St,
so there's no point in caching in Muffin. In a future PR, the atom
origin state should be detangled from the icon cache struct.