window: Remove some window icon properties and related processing (#399)
* window: Remove window icon properties and related processing

This isn't used in Cinnamon anymore, except for the notify::icon signal
which can be replaced with icon-changed.

* Add back the icon creation functions, fix blurry icons

Removing these prevents us from getting icons for some window backed
apps, such as Melange - which could solve this by using XAppGtkWindow,
but this could pose an issue for third party apps that have window
backed windows.

To fix the blurry icons, we are going to convert
meta_window_update_icon_now to meta_window_set_icon, which will accept
any size that will be passed in from `window_backed_app_get_icon` in
Cinnamon. We are not going to create icons for every window anymore, and
instead let Cinnamon use this function selectively when it needs to.

We can still remove the mini icon, and all of the theme/frame processing
because we don't set the icon in the title bar.

* Fix OR case in set_icon

* Don't warn when the icon is NULL

Fallback behavior was moved to Cinnamon, so there will be legitimate
cases when this is NULL.

* Rename meta_window_set_icon to meta_window_create_icon, add docs

* Always clear the icon cache

Now that icons are initialized from Cinnamon, we need to ensure an icon
is always created for the specified size. The texture is copied to St,
so there's no point in caching in Muffin. In a future PR, the atom
origin state should be detangled from the icon cache struct.