window-actor: Fix partially visible frameless windows, some window flickering cases (#405)
* window-actor: Fix partially visible frameless windows



Don't create mask textures for undecorated windows if there's no shape region

There was a mistake early on when moving mask generation invocation to
MetaWindowActor - which was clearing the mask texture for windows with
no shape region. This led to the other workarounds dealing with window
artifacts on size change because the mask was being created for windows
that didn't need it.

This commit has been tested for all cases that previously caused issues:

- Corners not updating on size/position change
- Chromium window corners with white artifacts
- Frameless windows being partially visible
- Xpad artifacts occurring when toggling decorations while effects are
enabled

* window-actor: Remove window group actor flag and related workaround

* Toggle the NO_LAYOUT actor flag on window_group before and after grab operations

Adding this flag to the window group was originally intended to help
improve window dragging smoothness, but its benefit there seems
negligible. However, it does improve input latency with window
rendering, so it's worth
finding a way to make it work without artifacts.

This disables the NO_LAYOUT flag during window dragging, which was
causing windows to flicker in and out of view.

* Fix build