Always get an accurate MetaFrameControl result
This fixes incorrect cursor icons when entering the title bar area.

While using coordinates from the x event is fast, it's not always
accurate, so do a round-trip and get the coordinates again IF the cursor
is not inside the client area (user is interacting with the
application). Our concerns here are only with frame operations, which
are not applicable in that case.