window: Cache the outer, client rects (#427)
* window: Cache the outer, client rects

* Replace meta_window_get_outer_rect with pointer access

Even though the value is now cached, this is one less function that
needs to be called in performance sensitive areas.

* Fix fullscreen CSD windows failing the fullscreen check

* Fix artifacts when some clients resize

Steam does not have a shape region from atom data, so when
check_needs_reshape is called, we're basing its fake shape region on the
client area, which at that point has stale coordinates. To fix this,
always refresh the rect caches when the actor geometry is synchronized.