Distinguish between internal and external (JS) unobscured overrides
When meta_window_actor_set_obscured_timed is set to false, the
obscured_lock override is in effect until it is called again with true,
which then waits 250ms and calls check_obscured after disabling the
override. This allows the windows underneath a closing window to start
painting before the user sees them. The timeout method is only used
internally in muffin for known cases like this.

This was causing a race condition though with GWL, where the thumbnail
menu would be open while a window is mapped, which would reset the
override that the thumbnail menu set for its set of windows. I
experimented with using obscured_lock as a counter, like freeze_count
works, but that made things too complicated. This adds a seperate state
for public overrides originating from JS so it and internal muffin
methods don't interact with each other's handling of overriding the
optimization.

Fix codacy warnings