Improve sync between the cursor and frames (#397)
* frames: Only skip updating prelight when both control and state match

prelit_control is used for both prelight and pressed states, so the early
return in update_prelit_control() misses the case where prelit_control
already matches the control we are updating, but its state is PRESSED
rather than PRELIGHT. Fix the condition to not have pressed controls
linger around erroneously.

https://bugzilla.gnome.org/show_bug.cgi?id=731058

* frames: Make sure to initialize button_state

Otherwise, we're comparing with uninitialized memory. Spotted by
valgrind.

* frames: Add MetaWindow pointer

* ui: Don't use grab ops for grab negotiation

https://github.com/gnome/mutter/commit/ebb6847bd11a1f56af5678d94228b28a522da2a5

* frames: Keep information about the ongoing grab operation, and retry if needed.

https://github.com/GNOME/mutter/commit/d05b750b8dec3825d286e3ae30c41b41bd6aed98

* frames: Don't make a round trip to the X server for every MotionNotify

https://github.com/GNOME/mutter/commit/799de4f0f48fc4d3f67df9f4ad2b31baf5296d91

* Remove a few more guards for frees

https://github.com/GNOME/mutter/commit/cd4206764e716f8df8390fa1ab39ee05db3d1835

* ui/frames: Simplify client area control computation

https://github.com/GNOME/mutter/commit/1f20e82a96eeb63207de9849b007a8da9f3002b5

* Protect against stray calls to meta_display_end_grab_op()

https://github.com/gnome/mutter/commit/ae7aabd5de134c48ec8cad9ecb15b06f57b16fa1

* frames: Only calculate frame geometry when the window rect size changes

* frames: Don't process motion events without the cursor

* frames: Fix cursor not switching from a resize state when hovering over the title bar or buttons

* Give tiled windows an extra pass on frame geometry calculation

Quarter tiled windows have artifacts when the frame geometry only
updates on size change, so let windows with a tile mode active receive
another pass as needed.

* Re-calculate frame geometry on move/resize

* Only call meta_frames_calc_geometry on frame manage and size change

Before, functions received MetaFrameGeometry by re-calculating it every
time, but it really will never change unless the window rect changes, as
that is what is passed to meta_theme_calc_geometry. It is an expensive
function.

The first attempt at optimizing this was checking the rect whenever the
calculation function was called, and storing a single MetaFrameGeometry
object on the MetaFrames struct, and returning the cache on no size
change. That was incorrect because MetaFrameGeometry is shared by all
windows. As a result, artifacts were occurring. When simply moving the
cache to MetaUIFrame, this caused an increase in latency.

The new approach is to only call meta_frames_calc_geometry in response
to size change, and to replace previous places where
meta_frames_calc_geometry is called with the cached object on
MetaUIFrame. This ensures the geometry is only calculated when it needs
to be, and that it stays current.

* Store a display pointer on MetaFrames

Reduce function invocation

* Remove unused last_motion_frame

* Avoid using GDK for fetching pointer coordinates

I have noticed gdk_window_get_device_position is a heavy function call,
so let's use some direct methods that provide the same functionality
with less abstraction.

The GDK method being replaced does not exist in GTK 4.

* Always get an accurate MetaFrameControl result

This fixes incorrect cursor icons when entering the title bar area.

While using coordinates from the x event is fast, it's not always
accurate, so do a round-trip and get the coordinates again IF the cursor
is not inside the client area (user is interacting with the
application). Our concerns here are only with frame operations, which
are not applicable in that case.

* frames: Ignore events for CSD windows

This fixes the cursor icon issue that has been affecting this branch without needing to use a round trip to query the pointer.

* frames: Use window->client_area

Been applying this commit locally for a while because it depended on #427, but this has a noticeable performance affect because get_control is called a lot.

* ui: Remove unused function meta_ui_unmap_frame