Simplify effects in progress state
These seperate counters were needed before effects composition was moved
to Cinnamon. Since meta_window_actor_effects_in_progress will return
true if any of the counters are greater than 1, they can be combined
into one. This allows us to kill meta_window_actor_effects_in_progress,
which is called during a lot of window operations.

The only other use for separate counters is distinguishing the effect
during debugging, but since this is a dev feature, we might as well just
pass the event number, and it can be looked up in the rare event the
effects counter is less than 0.