Rework map effect handling
Still not a complete fix, but seems to fix all of the issues except in
mintupdate, which still has an artifact as a result of it changing its
size/position mid-animation, but corrects itself at the end of the map
effect.

My current theory is this is happening because MetaWindowActor is being
created before MetaWindow receives the initial size hints from the
client (mintupdate).