Fix window mapping artifacts (WIP)
We shouldn't have to update the clutter actor after the texture is set.
After testing reverts of #427 and #397, the problem persists. This is
occurring because of how MetaWindowActor and MetaShapedTexture were
reconciled. I'm not happy with this, but it will at least minimize the
problem while trying to find a proper fix.

This also replaces the placement heuristic left over from 7+ years ago
with something a little more predictable and sane.