Fix mapping bugs
A lot of things are moving. Breakdown:

- clutter_actor_queue_relayout override fixes the position jumping when
mintupdate maps without breaking the animation. When initially mapping
we immediately queue a relayout on the container, and then subsequently
revert to normal behavior.

- The workarounds in meta_window_actor_effect_completed and
first_frame_drawn are removed.

- Size is only being set in sync_actor_geometry during the initial
mapping because it preempts the paint cycle, for all other cases we now
set it in set_cogl_texture, once the texture based on the window or
effect is generated.