window-actor: override Clutter opacity property and method (#449)
* window-actor: override Clutter opacity property and method

This restores actor opacity management without relying on a signal.

* Simplify effects in progress state

These separate counters were needed before effects composition was moved
to Cinnamon. Since meta_window_actor_effects_in_progress will return
true if any of the counters are greater than 1, they can be combined
into one. This allows us to kill meta_window_actor_effects_in_progress,
which is called during a lot of window operations.

The only other use for separate counters is distinguishing the effect
during debugging, but since this is a dev feature, we might as well just
pass the event number, and it can be looked up in the rare event the
effects counter is less than 0.

* Move MetaWindowActorPrivate struct to meta-window-actor-private.h

This was part of a local branch I wasn't initially going to include here, but it will make merging easier later.

* Never stop painting windows with pending effects

* window-actor: Respect client opacity

* Fix unminimizing windows

* Fix unsigned integer comparison

* Avoid wrapper usage

One less function call.

* Clean up

This wasn't needed.

* Remove unused function meta_prop_get_cardinal

* Fix usage of negative guint8 value