compositor: Fix workspace switching, improve window dragging smoothness (#465)
* compositor: Improve window dragging smoothness

* Undo the optimization disabling exemption

The net effect of having to check this on an interval negates the
benefit of the optimization being disabled.

* Don't update the client area with the outer rect

The client area rect is new since 4.0, and is only used in
MetaWindowActor and MetaFrames. The frame size change case covers
MetaFrames, and the call to update_client_area_rect covers the case in
process_shape_region.

* Store pointer to client_area instead

This should result in less allocation.

* Update outer rect in meta_window_show

* window-actor: Clean up

- Inlines some methods in perf-sensitive areas
- Now that we have the top_window_actor cached from the unredirection fix PR, we don't need to call g_list_last on compositor->windows, which wasn't correct.
- Removes dead code.

* window-group: Improve paint performance

Since the compositor now has access to MetaWindowActorPrivate, we can
get rid of some functions calls and add some equality checks before
copying regions.

* Revert conditional flag setting

* Remove is_move_resize

The idea was to check if it was only a move or resize grab, not a
compositor grab, but the function call negates the rendering benefit of
clutter_paint_volume_set_from_allocation.

* Fix rendering issues during and after workspaces switching

* Fix codacy nit

* Revert relayout change

* Restore workspace switch animation fix

Checking size_changed and position_changed in queue_relayout is wrong
because those states do not reset until the next redraw, and Clutter can
queue multiple relayouts in the interim. It's this additional window
rendering that is breaking the workspace switch animation in some
setups.