window.c: Remove a condition for updating the outer rect
We only update the outer rect if the edge resistance setting is enabled. Since
this setting is on by default it's being called on my most users installs
anyway. This also has the side effect of fixing other things that depend on
the rect being updated. An example of this is the workspace switcher Cinnamon
applet.