Fix screensaver regressions (#489)
* Fix windows not rendering after the screensaver is deactivated

* Fix the screensaver not updating

* Clarify timeout values

* Make sure the window group paints as the screensaver maps

* Revert "Make sure the window group paints as the screensaver maps"

This reverts commit a6a4f11fc0c49ae1f1ff834b03f8ef798c25dbcd.

* Revert "Fix the screensaver not updating"

This reverts commit 6d02afc857ba52dc7dd8c3bb5a383f95cae59413.

* MetaWindowGroup: fix paint volume

In the past, MetaWindowGroup was allocated the size of the screen and
painted the size of the screen because it contained the screen
background,
but now we also have the "top window group" which contains only popup
windows, so the allocation doesn't properly reflect the paint bounds
of the window group. Compute the paint bounds accurately from the
children.

https://bugzilla.gnome.org/show_bug.cgi?id=719669

https://github.com/GNOME/mutter/commit/a796938b3991535ac9b5e79436ce1dbf1da72904

* MetaWindowGroup: Only compensate for transforms when in clone paint

We were compensating for a clone paint viewport offset even when we
were not in clone paniting mode. This would break painting if we offset
the viewport for some other reason for example as in the future stage
view painting.

https://bugzilla.gnome.org/show_bug.cgi?id=768976

https://github.com/GNOME/mutter/commit/749237a28e87a46a8b0fbe0415628c99b406a54f