Rebase clean up step 2: move MetaWindowActorPrivate to meta-window-actor-private.h
We need to move the MetaWindowActorPrivate struct back to the class
private header file though because that's how it's accessed in the
original patch. I don't want to make life for myself more difficult in
the future, or in compositor.c call a newly created method in
meta-window-actor.c in busy rendering code paths. MetaCompositor tracks
and manages all actor windows, so it should have access to their private
state. This change has no negative side effects, is forward compatible
with the obscured work slated for 4.4, and just gives MetaCompositor
full scope for the class instances it manages.