window.c, workspace.c: Remove delayed emission of the 'window-added' signals.
This caused complications with signals being fired out of order,
resulting in window list issues (having the wrong window list
tracking a window, for example).  Also, the window argument in the
callback could only ever point to a valid MetaWindow, or some
undefined memory location, so checking for null was pointless.

This reverts commit c3ffb540e969b9a1d938d83b4662f55a1cf4278f.